requirements.txt
# Core dependencies for Azure RAG POC
openai==1.12.0
langchain==0.1.6
langchain-openai==0.0.5
faiss-cpu==1.7.4
tiktoken==0.5.2
python-dotenv==1.0.0

# Document processing
pypdf==4.0.1
requests==2.31.0
beautifulsoup4==4.12.2

# Scientific computing
numpy==1.24.3

# Jupyter
jupyterlab==4.0.11
ipykernel==6.29.0

# Optional: For enhanced document parsing
# python-docx==1.1.0
# openpyxl==3.1.2

